pyproject.toml
[tool.poetry]
name = "simpkcs12"
version = "0.1.0"
description = "Simple PKCS12 tools."
authors = [
    "João Paulo Taylor Ienczak Zanette <jpaulotiz@gmail.com>",
    "Larissa Taw Rumiana de Oliveira <>",
]

[tool.poetry.dependencies]
carl = "*"
python = "^3.7"
pyOpenSSL = "^19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
